api/requirements.txt
flask==2.3.2
prometheus-flask-exporter==0.23.1
kafka-python==2.0.2